Moved common code for files to fs/file_util
